refactor: useArticles 이용해서 ariticle 데이터 불러오기
