+++ client/src/App.tsx
@@ -1,8 +1,15 @@
 import { RouterProvider } from "react-router-dom";
 import routes from "./routers/routes";
+import { QueryClient, QueryClientProvider } from "react-query";
+
+const queryClient = new QueryClient();
 
 function App() {
-  return <RouterProvider router={routes} />;
+  return (
+    <QueryClientProvider client={queryClient}>
+      <RouterProvider router={routes} />
+    </QueryClientProvider>
+  );
 }
 
 export default App;
